+++ app/modules/delay.js
@@ -1,0 +1,7 @@
+export function delay(interval = 300) {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve();
+      }, interval);
+    });
+}
