+++ app/modules/sanitizeHtml.js
@@ -1,0 +1,7 @@
+export function secureReplace(string) {
+    return string
+    .replaceAll("&", "&amp;")
+    .replaceAll("<", "&lt;")
+    .replaceAll(">", "&gt;")
+    .replaceAll('"', "&quot;")
+}
