+++ app/modules/tryPostAgain.js
@@ -1,0 +1,8 @@
+export function tryPostAgain(error, func) {
+    if (error.message === 'Сервер упал') {
+        func()
+    //В ином случае алертим сообщение об ошибке
+    } else {
+        alert(error)
+    }
+}
