+++ index.html
@@ -3,18 +3,18 @@
   <head>
     <title>Проект "Комменты"</title>
     <meta charset="utf-8" />
-    <link rel="stylesheet" href="styles.css" />
-    <script defer src="./app.js"></script>
+    <link rel="stylesheet" href="./css/styles.css" />
+    <script type="module" defer src="./app/app.js"></script>
   </head>
 
   <body>
     
     <div class="container">
-      <div id="loading-head" class="loading loading_hidden"><img class="loading-img" width="300px" src="./loading-head.gif" alt="loading-bar">Комментарии загружаются</div>
+      <div id="loading-head" class="loading loading_hidden"><img class="loading-img" width="300px" src="./img/loading-head.gif" alt="loading-bar">Комментарии загружаются</div>
       <ul id="comments-box" class="comments">
         <!-- Рендерим список из JS -->
       </ul>
-      <div id="loading-comments" class="loading loading_hidden"><img class="loading-img" width="300px" src="./loading-head.gif" alt="loading-bar">Комментарий загружается</div>
+      <div id="loading-comments" class="loading loading_hidden"><img class="loading-img" width="300px" src="./img/loading-head.gif" alt="loading-bar">Комментарий загружается</div>
       <div class="add-form">
         <input
           id="name-input"
